refactor: Price VO의 package 변경
