+++ apps/commerce-api/src/main/java/com/loopers/domain/product/ProductSummary.java
@@ -1,5 +1,6 @@
 package com.loopers.domain.product;
 
+import com.loopers.domain.vo.Price;
 import java.time.LocalDate;
 
 public record ProductSummary(
