+++ apps/commerce-api/src/main/java/com/loopers/domain/vo/Price.java
@@ -1,4 +1,4 @@
-package com.loopers.domain.product;
+package com.loopers.domain.vo;
 
 import com.loopers.support.error.CoreException;
 import com.loopers.support.error.ErrorType;
