+++ apps/commerce-api/src/test/java/com/loopers/domain/like/LikeIntegrationTest.java
@@ -11,11 +11,11 @@ import static org.mockito.Mockito.verify;
 import com.loopers.application.like.LikeFacade;
 import com.loopers.application.like.LikeProductCommand;
 import com.loopers.domain.brand.BrandEntity;
-import com.loopers.domain.product.Price;
 import com.loopers.domain.product.ProductEntity;
 import com.loopers.domain.product.Stock;
 import com.loopers.domain.user.Gender;
 import com.loopers.domain.user.UserEntity;
+import com.loopers.domain.vo.Price;
 import com.loopers.infrastructure.brand.BrandJpaRepository;
 import com.loopers.infrastructure.product.ProductJpaRepository;
 import com.loopers.infrastructure.user.UserJpaRepository;
