+++ apps/commerce-api/src/test/java/com/loopers/domain/order/OrderIntegrationTest.java
@@ -15,11 +15,11 @@ import com.loopers.application.order.OrderFacade;
 import com.loopers.application.order.OrderInfo;
 import com.loopers.application.order.PaymentInfo;
 import com.loopers.domain.brand.BrandEntity;
-import com.loopers.domain.product.Price;
 import com.loopers.domain.product.ProductEntity;
 import com.loopers.domain.product.Stock;
 import com.loopers.domain.user.Gender;
 import com.loopers.domain.user.UserEntity;
+import com.loopers.domain.vo.Price;
 import com.loopers.infrastructure.brand.BrandJpaRepository;
 import com.loopers.infrastructure.order.OrderJpaRepository;
 import com.loopers.infrastructure.product.ProductJpaRepository;
